Make saveData
Create a function that generates multiple arrays of different sizes and
creates a set of data containing the time taken for a callback to run on
each array of different size.

Save the data points in pairs with the x point being the length of the array
and the y point being the time taken for the callback to run on the array
of that length.