Finish shuffle plotting
Make the necessary files and code for shuffle in the same way as the
other functions.

Modify singleTimer() and saveData() to take another
paramter which is the number of data points to take an average from.
This is so that the code is able to adjust for functions of different
speeds.