Make general solution
Modify subSequenceSum() to account for any array of any length with
any target given. Do this by using a for loop at each index of the array
and adding each subsequent element one by one and checking if the sum
equals the target value.

If the sum does equal the target, simply return true which will break
out of the loop. In the event where this condition is not met, the
loops will simply end and the function will get to the end line and
just return false.