Restructure
Modify saveFibData() to use singleTimer() instead of its own version by
simply changing the first parameted passed into an integer rather
than an array. This means singleFibTimer() can its test file can be
deleted.

Also, modify singleTimer() and singleData() to pass in a random number
generated which will be the target for subSequenceTimer(). This will not
affect the other custom algorithms as they do not take a second argument
and so the target parameter passed into them will not make a difference.