Update shuffle
Modify shuffle to not use sort() in order to limit the number of inbuilt
functions used for custom made algorithms.

Instead, create a new array and randomly select an element, one at a time,
from the original array and move it into the new array, which should shuffle
all the elements. Simply return this array at the end of the function.